move all pages of AssignTask panel to AssignTaskPages folder, rename wrong names of pages in AssignTaskPages
