+++ CuoiKi/CuoiKi.csproj.user
@@ -10,10 +10,10 @@
     <Compile Update="UI\Forms\StageForm.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Update="UI\Manager\AssignTaskForm.xaml.cs">
+    <Compile Update="UI\Manager\AssignTaskPages\ProjectsPage.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Update="UI\Manager\AssignTaskPages\ProjectPage.xaml.cs">
+    <Compile Update="UI\Manager\AssignTaskPages\StagesPage.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
     <Compile Update="UI\Manager\CalcSalaryForm.xaml.cs">
@@ -48,10 +48,10 @@
     <Page Update="UI\Forms\StageForm.xaml">
       <SubType>Designer</SubType>
     </Page>
-    <Page Update="UI\Manager\AssignTaskForm.xaml">
+    <Page Update="UI\Manager\AssignTaskPages\ProjectsPage.xaml">
       <SubType>Designer</SubType>
     </Page>
-    <Page Update="UI\Manager\AssignTaskPages\ProjectPage.xaml">
+    <Page Update="UI\Manager\AssignTaskPages\StagesPage.xaml">
       <SubType>Designer</SubType>
     </Page>
     <Page Update="UI\Manager\CalcSalaryForm.xaml">
