+++ CuoiKi/UI/Manager/AssignTaskPages/StagesPage.xaml
@@ -1,14 +1,15 @@
-﻿<Page x:Class="CuoiKi.UI.Manager.AssignTaskForm"
+﻿<Page x:Class="CuoiKi.UI.Manager.AssignTaskPages.StagesPage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:Icon="http://metro.mahapps.com/winfx/xaml/iconpacks"
       xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:CuoiKi.UI.Manager"
+      xmlns:local="clr-namespace:CuoiKi.UI.Manager.AssignTaskPages"
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="800"
-      Title="AssignTaskForm">
+      Title="ProjectPage">
 
-    <Grid >
+    <Grid>
         <Grid.RowDefinitions>
             <RowDefinition Height="*"></RowDefinition>
             <RowDefinition Height="2*"></RowDefinition>
@@ -19,9 +20,9 @@
             <ColumnDefinition Width="2*"></ColumnDefinition>
             <ColumnDefinition Width="*"></ColumnDefinition>
         </Grid.ColumnDefinitions>
-        <Label Content="Projects" FontSize="36" Grid.ColumnSpan="2" HorizontalAlignment="Center" VerticalAlignment="Center"/>
-        <Button Grid.Column="3" Content="Add" HorizontalAlignment="Center" VerticalAlignment="Center"/>
-        <ScrollViewer Width="800" Height="350" Grid.Row="1" Grid.Column="0" Grid.RowSpan="2" Grid.ColumnSpan="3" x:Name="MyScrollViewer" HorizontalScrollBarVisibility="Visible" VerticalScrollBarVisibility="Visible">
+        <Label Content="Stages" FontSize="36" Grid.ColumnSpan="2" HorizontalAlignment="Center" VerticalAlignment="Center"/>
+        <Button x:Name="BtnAddStage" Grid.Column="3" Content="Add" HorizontalAlignment="Center" VerticalAlignment="Center" Click="BtnAddStage_Click"/>
+        <ScrollViewer Width="800" Height="350" Grid.Row="2" Grid.Column="0" Grid.RowSpan="2" Grid.ColumnSpan="5" x:Name="MyScrollViewer" HorizontalScrollBarVisibility="Visible" VerticalScrollBarVisibility="Visible" Margin="0,5,0,5">
             <ItemsControl ItemsSource="{Binding TestList}">
                 <ItemsControl.ItemsPanel>
                     <ItemsPanelTemplate>
@@ -30,15 +31,22 @@
                 </ItemsControl.ItemsPanel>
                 <ItemsControl.ItemTemplate>
                     <DataTemplate>
-                        <Button Margin="10 20 10 20" Width="200" Height="100" Click="btn_Click">
+                        <Button Margin="10 20 10 20" Width="200" Height="100" Click="BtnAddStage_Click">
                             <StackPanel Orientation="Vertical">
-                                <TextBlock Text="{Binding ID}" Margin="5,0"></TextBlock>
-                                <TextBlock Text="{Binding Name}" Margin="5,0"></TextBlock>
+                                <TextBlock Text="{Binding StageName}" Margin="5,0"></TextBlock>
+                                <TextBlock Text="{Binding LeaderName}" Margin="5,0"></TextBlock>
                             </StackPanel>
                         </Button>
                     </DataTemplate>
                 </ItemsControl.ItemTemplate>
             </ItemsControl>
         </ScrollViewer>
+        <Button x:Name="Back_click" Grid.Column="0"
+                VerticalAlignment="Center" 
+                HorizontalAlignment="Left"
+                Margin="40 0 0 0" Click="back_click">
+            <TextBlock Text="←" FontSize="30" Margin="0 -10 0 0"/>
+        </Button>
+
     </Grid>
 </Page>
